client: add support for uniview:// urls
To work with the planned interface plans, it is necessary to open a room
programmatically. The easiest way to do that is by adding a custom
protocol and define the program as its handler. This has the added
advantage that it also allows for a potential third party program to
utilise a uniview server.